Refactor and alter tests
