+++ cohort-cli/src/main/java/com/ibm/cohort/cli/CohortCLI.java
@@ -31,13 +31,13 @@ import com.ibm.cohort.engine.CqlEvaluator;
 import com.ibm.cohort.engine.DirectoryLibrarySourceProvider;
 import com.ibm.cohort.engine.EvaluationResultCallback;
 import com.ibm.cohort.engine.FhirLibraryLibrarySourceProvider;
-import com.ibm.cohort.engine.LibraryFormat;
 import com.ibm.cohort.engine.LoggingEnum;
 import com.ibm.cohort.engine.MultiFormatLibrarySourceProvider;
 import com.ibm.cohort.engine.TranslatingLibraryLoader;
 import com.ibm.cohort.engine.ZipStreamLibrarySourceProvider;
 import com.ibm.cohort.engine.helpers.FileHelpers;
 import com.ibm.cohort.fhir.client.config.FhirClientBuilderFactory;
+import com.ibm.cohort.file.LibraryFormat;
 import com.ibm.cohort.translator.provider.CqlTranslationProvider;
 import com.ibm.cohort.translator.provider.InJVMCqlTranslationProvider;
