+++ cohort-engine/src/main/java/com/ibm/cohort/engine/ZipStreamLibrarySourceProvider.java
@@ -20,6 +20,7 @@ import org.hl7.elm.r1.VersionedIdentifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.ibm.cohort.file.LibraryFormat;
 import com.ibm.cohort.version.DefaultFilenameToVersionedIdentifierStrategy;
 import com.ibm.cohort.version.FilenameToVersionedIdentifierStrategy;
