+++ cohort-translation/src/main/java/com/ibm/cohort/translator/provider/BaseCqlTranslationProvider.java
@@ -10,6 +10,10 @@ import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
+
+import com.ibm.cohort.annotations.Generated;
+import com.ibm.cohort.file.LibraryFormat;
+
 import org.cqframework.cql.cql2elm.CqlTranslator.Options;
 import org.cqframework.cql.cql2elm.CqlTranslatorOptions;
 import org.cqframework.cql.elm.execution.Library;
@@ -18,6 +22,7 @@ import org.cqframework.cql.elm.execution.Library;
  * Common functionality for use when implementing CQL translator
  * wrapper implementations.
  */
+@Generated
 public abstract class BaseCqlTranslationProvider implements CqlTranslationProvider {
 
 	public static final LibraryFormat DEFAULT_TARGET_FORMAT = LibraryFormat.XML;
